requirements.txt


Flask==2.2.2
Werkzeug==2.3.6
numpy
opencv-python
scikit-image
pillow

